Support #20355 Implement initial specs of Collecting Event Imperfect but working solution to expose partial dates as String and store them as datetim with defined precisions. More work required to make it more robust and cleaner.

+++ src/test/java/ca/gc/aafc/collection/api/entities/CollectingEventCRUDIT.java
@@ -34,7 +34,7 @@ public class CollectingEventCRUDIT extends CollectionModuleBaseIT {
         .decimalLatitude(12.123456)
         .decimalLongitude(45.01)
         .startEventDateTime(testDateTime)
-        .startEventDateTimePrecision((short) 8)
+        .startEventDateTimePrecision((byte) 8)
         .build();
     dbService.save(collectingEvent);
 
@@ -43,7 +43,7 @@ public class CollectingEventCRUDIT extends CollectionModuleBaseIT {
     assertEquals(12.123456, fetchedCollectingEvent.getDecimalLatitude());
     assertEquals(45.01, fetchedCollectingEvent.getDecimalLongitude());
     assertEquals(testDateTime, fetchedCollectingEvent.getStartEventDateTime());
-    assertEquals((short) 8, fetchedCollectingEvent.getStartEventDateTimePrecision());
+    assertEquals((byte) 8, fetchedCollectingEvent.getStartEventDateTimePrecision());
     assertNotNull(fetchedCollectingEvent.getCreatedOn());
 
   }
